+++ src/AddSkillReviewForm.test.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import SkillReviewsForm from './SkillReviewsForm';
+import AddSkillReviewForm from './AddSkillReviewForm';
 import { shallow, mount } from 'enzyme';
 
-describe('<SkillReviewsForm />', () => {
+describe('<AddSkillReviewForm />', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<SkillReviewsForm />, div);
+    ReactDOM.render(<AddSkillReviewForm />, div);
   });
 
   it('matches the stored snapshot', () => {
-    const wrapper = shallow(<SkillReviewsForm />);
+    const wrapper = shallow(<AddSkillReviewForm />);
     expect(wrapper.get(0)).toMatchSnapshot();
   });
 
   it('changes state when input to "Team Member" field changes', () => {
-    const wrapper = mount(<SkillReviewsForm />);
+    const wrapper = mount(<AddSkillReviewForm />);
     expect(wrapper.state().team_member_id).toBe('');
     const newID = '1234';
     const event = { target: { value: newID } };
@@ -24,7 +24,7 @@ describe('<SkillReviewsForm />', () => {
   });
 
   it('changes state when input to "Positive Review" field changes', () => {
-    const wrapper = mount(<SkillReviewsForm />);
+    const wrapper = mount(<AddSkillReviewForm />);
     expect(wrapper.state().positive).toBe(true);
     const newVal = false;
     const event = { target: { value: newVal } };
@@ -33,7 +33,7 @@ describe('<SkillReviewsForm />', () => {
   });
 
   it('should be accept Strings as ev.target.value', () => {
-    const wrapper = mount(<SkillReviewsForm />);
+    const wrapper = mount(<AddSkillReviewForm />);
     const positiveSelectElement = wrapper.find('[name="positiveSelect"]');
     const falseEvent = {
       target: {
@@ -52,7 +52,7 @@ describe('<SkillReviewsForm />', () => {
   });
 
   it('changes state when text in "Review" text area changes', () => {
-    const wrapper = mount(<SkillReviewsForm />);
+    const wrapper = mount(<AddSkillReviewForm />);
     const newBody = 'Little Bo Peep';
     expect(wrapper.state().body).toBe('');
     const event = { target: { value: newBody } };
